LPS-63229 Move to json package like com.liferay.portal.json.JSONInit
